Revert changes 09d8b08..=3d3da02
This commit effectively returns HEAD to commit `8c61af2`.
At this point, these were the changes to the original fuzz target:

Instead of using only the default options for the `PrettyPrinter`,
randomize the options by building a `PrettyPrinter` instance
using the fuzz input, and feeding the created `PrettyPrinter` an
arbitrary string
